# docs/api/openapi.yml
components:
  schemas:
...
    ErrorsJSON:
      properties:
        code:
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorsResponseerroradditional'
          type: array
        message:
          type: string
        url:
          type: string
      required:
      - code
      - message
      type: object
    ErrorsResponseerroradditional:
      properties:
        message:
          type: string
      type: object
...
    Querybuildertypesv5ClickHouseQuery:
      properties:
        disabled:
          type: boolean
        legend:
          type: string
        name:
          type: string
        query:
          type: string
      type: object
...
    Querybuildertypesv5CompositeQuery:
      description: Composite query containing one or more query envelopes. Each query
        envelope specifies its type and corresponding spec.
      properties:
        queries:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelope'
          nullable: true
          type: array
      type: object
...
    Querybuildertypesv5Filter:
      properties:
        expression:
          type: string
      type: object
    Querybuildertypesv5FormatOptions:
      properties:
        fillGaps:
          type: boolean
        formatTableResultForUI:
          type: boolean
      type: object
    Querybuildertypesv5Function:
      properties:
        args:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5FunctionArg'
          type: array
        name:
          $ref: '#/components/schemas/Querybuildertypesv5FunctionName'
      type: object
    Querybuildertypesv5FunctionArg:
      properties:
        name:
          type: string
        value: {}
      type: object
    Querybuildertypesv5FunctionName:
      enum:
      - cutoffmin
      - cutoffmax
      - clampmin
      - clampmax
      - absolute
      - runningdiff
      - log2
      - log10
      - cumulativesum
      - ewma3
      - ewma5
      - ewma7
      - median3
      - median5
      - median7
      - timeshift
      - anomaly
      - fillzero
      type: string
    Querybuildertypesv5GroupByKey:
      properties:
        description:
          type: string
        fieldContext:
          $ref: '#/components/schemas/TelemetrytypesFieldContext'
        fieldDataType:
          $ref: '#/components/schemas/TelemetrytypesFieldDataType'
        name:
          type: string
        signal:
          $ref: '#/components/schemas/TelemetrytypesSignal'
        unit:
          type: string
      required:
      - name
      type: object
    Querybuildertypesv5Having:
      properties:
        expression:
          type: string
      type: object
...
    Querybuildertypesv5OrderBy:
      properties:
        direction:
          $ref: '#/components/schemas/Querybuildertypesv5OrderDirection'
        key:
          $ref: '#/components/schemas/Querybuildertypesv5OrderByKey'
      type: object
    Querybuildertypesv5OrderByKey:
      properties:
        description:
          type: string
        fieldContext:
          $ref: '#/components/schemas/TelemetrytypesFieldContext'
        fieldDataType:
          $ref: '#/components/schemas/TelemetrytypesFieldDataType'
        name:
          type: string
        signal:
          $ref: '#/components/schemas/TelemetrytypesSignal'
        unit:
          type: string
      required:
      - name
      type: object
    Querybuildertypesv5OrderDirection:
      enum:
      - asc
      - desc
      type: string
    Querybuildertypesv5PromQuery:
      properties:
        disabled:
          type: boolean
        legend:
          type: string
        name:
          type: string
        query:
          type: string
        stats:
          type: boolean
        step:
          $ref: '#/components/schemas/Querybuildertypesv5Step'
      type: object
    Querybuildertypesv5QueryBuilderFormula:
      properties:
        disabled:
          type: boolean
        expression:
          type: string
        functions:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5Function'
          type: array
        having:
          $ref: '#/components/schemas/Querybuildertypesv5Having'
        legend:
          type: string
        limit:
          type: integer
        name:
          type: string
        order:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5OrderBy'
          type: array
      type: object
...
    Querybuildertypesv5QueryBuilderTraceOperator:
      properties:
        aggregations:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5TraceAggregation'
          type: array
        cursor:
          type: string
        disabled:
          type: boolean
        expression:
          type: string
        filter:
          $ref: '#/components/schemas/Querybuildertypesv5Filter'
        functions:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5Function'
          type: array
        groupBy:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5GroupByKey'
          type: array
        having:
          $ref: '#/components/schemas/Querybuildertypesv5Having'
        legend:
          type: string
        limit:
          type: integer
        name:
          type: string
        offset:
          type: integer
        order:
          items:
            $ref: '#/components/schemas/Querybuildertypesv5OrderBy'
          type: array
        returnSpansFrom:
          type: string
        selectFields:
          items:
            $ref: '#/components/schemas/TelemetrytypesTelemetryFieldKey'
          type: array
        stepInterval:
          $ref: '#/components/schemas/Querybuildertypesv5Step'
      type: object
...
    Querybuildertypesv5QueryEnvelope:
      oneOf:
      - $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelopeBuilderTrace'
      - $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelopeBuilderLog'
      - $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelopeBuilderMetric'
      - $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelopeFormula'
      - $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelopeTraceOperator'
      - $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelopePromQL'
      - $ref: '#/components/schemas/Querybuildertypesv5QueryEnvelopeClickHouseSQL'
      properties:
        spec: {}
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryEnvelopeBuilderLog:
      properties:
        spec:
          $ref: '#/components/schemas/Querybuildertypesv5QueryBuilderQueryGithubComHanzo O11yObservePkgTypesQuerybuildertypesQuerybuildertypesv5LogAggregation'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryEnvelopeBuilderMetric:
      properties:
        spec:
          $ref: '#/components/schemas/Querybuildertypesv5QueryBuilderQueryGithubComHanzo O11yObservePkgTypesQuerybuildertypesQuerybuildertypesv5MetricAggregation'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryEnvelopeBuilderTrace:
      properties:
        spec:
          $ref: '#/components/schemas/Querybuildertypesv5QueryBuilderQueryGithubComHanzo O11yObservePkgTypesQuerybuildertypesQuerybuildertypesv5TraceAggregation'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryEnvelopeClickHouseSQL:
      properties:
        spec:
          $ref: '#/components/schemas/Querybuildertypesv5ClickHouseQuery'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryEnvelopeFormula:
      properties:
        spec:
          $ref: '#/components/schemas/Querybuildertypesv5QueryBuilderFormula'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryEnvelopePromQL:
      properties:
        spec:
          $ref: '#/components/schemas/Querybuildertypesv5PromQuery'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryEnvelopeTraceOperator:
      properties:
        spec:
          $ref: '#/components/schemas/Querybuildertypesv5QueryBuilderTraceOperator'
        type:
          $ref: '#/components/schemas/Querybuildertypesv5QueryType'
      type: object
    Querybuildertypesv5QueryRangeRequest:
      description: Request body for the v5 query range endpoint. Supports builder
        queries (traces, logs, metrics), formulas, joins, trace operators, PromQL,
        and ClickHouse SQL queries.
      properties:
        compositeQuery:
          $ref: '#/components/schemas/Querybuildertypesv5CompositeQuery'
        end:
          minimum: 0
          type: integer
        formatOptions:
          $ref: '#/components/schemas/Querybuildertypesv5FormatOptions'
        noCache:
          type: boolean
        requestType:
          $ref: '#/components/schemas/Querybuildertypesv5RequestType'
        schemaVersion:
          type: string
        start:
          minimum: 0
          type: integer
        variables:
          additionalProperties:
            $ref: '#/components/schemas/Querybuildertypesv5VariableItem'
          type: object
      type: object
...
    Querybuildertypesv5QueryType:
      enum:
      - builder_query
      - builder_formula
      - builder_trace_operator
      - clickhouse_sql
      - promql
      type: string
...
    Querybuildertypesv5RequestType:
      enum:
      - scalar
      - time_series
      - raw
      - raw_stream
      - trace
      type: string
...
    Querybuildertypesv5Step:
      description: Step interval. Accepts a Go duration string (e.g., "60s", "1m",
        "1h") or a number representing seconds (e.g., 60).
      oneOf:
      - description: Duration string (e.g., "60s", "5m", "1h").
        example: 60s
        type: string
      - description: Duration in seconds.
        example: 60
        type: number
...
    Querybuildertypesv5TraceAggregation:
      properties:
        alias:
          type: string
        expression:
          type: string
      type: object
    Querybuildertypesv5VariableItem:
      properties:
        type:
          $ref: '#/components/schemas/Querybuildertypesv5VariableType'
        value: {}
      type: object
    Querybuildertypesv5VariableType:
      enum:
      - query
      - dynamic
      - custom
      - text
      type: string
    RenderErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorsJSON'
        status:
          type: string
      required:
      - status
      - error
      type: object
...
    TelemetrytypesFieldContext:
      enum:
      - metric
      - log
      - span
      - resource
      - attribute
      - body
      type: string
    TelemetrytypesFieldDataType:
      enum:
      - string
      - bool
      - float64
      - int64
      - number
      type: string
...
    TelemetrytypesSignal:
      enum:
      - traces
      - logs
      - metrics
      type: string
...
    TelemetrytypesTelemetryFieldKey:
      properties:
        description:
          type: string
        fieldContext:
          $ref: '#/components/schemas/TelemetrytypesFieldContext'
        fieldDataType:
          $ref: '#/components/schemas/TelemetrytypesFieldDataType'
        name:
          type: string
        signal:
          $ref: '#/components/schemas/TelemetrytypesSignal'
        unit:
          type: string
      required:
      - name
      type: object
...
  securitySchemes:
    api_key:
      description: API Keys
      in: header
      name: Hanzo O11y-Api-Key
      type: apiKey
    tokenizer:
      bearerFormat: Tokenizer
      description: Tokens generated by the tokenizer
      scheme: bearer
      type: http
info:
  description: OpenTelemetry-Native Logs, Metrics and Traces in a single pane
  title: Hanzo O11y
  version: ""
openapi: 3.0.3
paths:
...
  /api/v5/substitute_vars:
    post:
      deprecated: false
      description: Replace variables in a query
      operationId: ReplaceVariables
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Querybuildertypesv5QueryRangeRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Querybuildertypesv5QueryRangeRequest'
                  status:
                    type: string
                required:
                - status
                - data
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderErrorResponse'
          description: Internal Server Error
      security:
      - api_key:
        - VIEWER
      - tokenizer:
        - VIEWER
      summary: Replace variables
      tags:
      - querier